backend/requirements.txt
django>=4.2.4,<5
djangorestframework>=3.14.0
django-environ>=0.10.0
django-cors-headers>=4.2.0
djangorestframework-simplejwt>=5.2.2


ruff>=0.1.5 #format code python
